refactor: clean working directory
- 쓰지 않는 파일 삭제
- db 관련 디렉토리 모으기
